- CV: replaced commented logger by print in setup_cv_iter.py that is redirected to log file. - CV: added argument `log_dir` in preprocessing_job.sh to reflect change in preprocess_cv_folds_trecord_dataset.py - CV: added external argument in preprocess_cv_folds_trecord_dataset.py so logs produced by the Python script go to a log directory created to store logs. - CV: in create_folds_list_for_cv.py, added code to create yaml file for normalized CV iterations and also yaml file with CV filenames for each CV iteration. - Kepler Simulated Data: created script create_dataset_inj1_plnt.py that builds a TFRecord data set with only planets and inj1 TCEs. - CV: logs created in create_cv_tfrecord_dataset_from_tables.py are now saved in a log directory; predict directory is only created if there are shards for predict. - CV: in create_cv_folds_tables.py, log is now save to the root directory for the shard tables; moving from using TCE table to create the shards tables, and use instead the tfrecord table only. - CV: commented fields in config_cv_train.yaml related to normalization; `tfrec_dir` set to `null` since we now use `cv_folds` only. - CV: in add_tfrec_dataset_fps_to_config_file.py, setting up `datasets_fps` now is based on the new yaml file.

+++ src_cv/create_cv_dataset/create_cv_tfrecord_dataset_from_tables.py
@@ -15,18 +15,20 @@ from src_cv.utils_cv import create_shard_fold, create_table_shard_example_locati
 if __name__ == '__main__':
 
     # CV data directory; contains the TCE tables for each fold
-    data_dir = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/TESS/cv_tess_s1-s67_updated_labels_02-02-2024_1210')
+    data_dir = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/Kepler/Q1-Q17_DR25/tfrecords_kepler_q1q17dr25_obsplanets_siminj1_2-22-2024_1115/cv_keplerq1q17dr25_obsplanets_siminj1_data_02-22-2024_1155')
     # TFRecord source directory; non-normalized examples
-    src_tfrec_dir = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/TESS/tfrecords_tess_s1-s67_all_12-6-2023_1039_adddiffimg_perimgnormdiffimg_updatedlabels_2-2-2024')
+    src_tfrec_dir = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/Kepler/Q1-Q17_DR25/tfrecords_kepler_q1q17dr25_obsplanets_siminj1_2-22-2024_1115/tfrecords_kepler_q1q17dr25obs_planets_sim_inj1_2-22-2024_1115')
     # table that maps a TCE to a given TFRecord file in the source TFRecords
-    src_tfrec_tbl_fp = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/TESS/tfrecords_tess_s1-s67_all_12-6-2023_1039_adddiffimg_perimgnormdiffimg_updatedlabels_2-2-2024/shards_tbl.csv')
+    src_tfrec_tbl_fp = Path('/home6/msaragoc/work_dir/Kepler-TESS_exoplanet/data/tfrecords/Kepler/Q1-Q17_DR25/tfrecords_kepler_q1q17dr25_obsplanets_siminj1_2-22-2024_1115/tfrecords_kepler_q1q17dr25obs_planets_sim_inj1_2-22-2024_1115/shards_tbl.csv')
     n_processes = 10
 
     data_dir.mkdir(exist_ok=True)
 
     # set up logger
+    log_dir = data_dir / 'create_cv_folds_logs'
+    log_dir.mkdir(exist_ok=True)
     logger = logging.getLogger(name=f'create_cv_folds_shards')
-    logger_handler = logging.FileHandler(filename=data_dir / f'create_cv_folds_shards.log', mode='w')
+    logger_handler = logging.FileHandler(filename=log_dir / f'create_cv_folds_shards.log', mode='w')
     logger_formatter = logging.Formatter('%(asctime)s - %(message)s')
     logger.setLevel(logging.INFO)
     logger_handler.setFormatter(logger_formatter)
@@ -74,51 +76,41 @@ if __name__ == '__main__':
     # assert num_unit_folds <= num_folds
     folds_arr = np.arange(num_folds)
 
-    cv_folds_runs = [{'train': data_shards[:fold_i] + data_shards[fold_i + 1:],
-                      'test': [data_shards[fold_i]]} for fold_i in folds_arr]
-
-    np.save(data_dir / f'cv_folds_runs.npy', cv_folds_runs)
-
-    cv_iterations = []
-    for cv_iter in cv_folds_runs:
-        cv_iterations.append([cv_iter['train'], cv_iter['test']])
-    cv_iterations = pd.DataFrame(cv_iterations, columns=['train', 'test'])
-    cv_iterations.to_csv(data_dir / f'cv_folds_runs.csv')
-
     # create TFRecord for examples that are not part of the evaluation set
-    logger.info('Creating shards for examples not used for evaluation...')
-    dest_tfrec_dir_noteval = data_dir / 'tfrecords' / 'predict'
-    dest_tfrec_dir_noteval.mkdir(exist_ok=True)
-    tce_tbl_noteval_fp = Path(data_dir / f'examples_noteval.csv')
-    # create_shard_fold(tce_tbl_noteval_fp, dest_tfrec_dir_noteval, -1, src_tfrec_dir, src_tfrec_tbl, True)
-
     shard_tbls_dir = data_dir / 'shard_tables' / 'predict'  # directory with the TCE tables for each fold
-    shard_tbls_fps = sorted(list(shard_tbls_dir.iterdir()))
+    if shard_tbls_dir.exists():
+        shard_tbls_fps = sorted(list(shard_tbls_dir.iterdir()))
 
-    # destination directory for the predicted shards
-    dest_tfrec_dir = data_dir / 'tfrecords' / 'predict'
-    dest_tfrec_dir.mkdir(exist_ok=True, parents=True)
+        logger.info('Creating shards for examples not used for evaluation...')
+        dest_tfrec_dir_noteval = data_dir / 'tfrecords' / 'predict'
+        dest_tfrec_dir_noteval.mkdir(exist_ok=True)
+        tce_tbl_noteval_fp = Path(data_dir / f'examples_noteval.csv')
+        # create_shard_fold(tce_tbl_noteval_fp, dest_tfrec_dir_noteval, -1, src_tfrec_dir, src_tfrec_tbl, True)
 
-    pool = multiprocessing.Pool(processes=n_processes)
-    jobs = [(shard_tbl_fp, dest_tfrec_dir, fold_i, src_tfrec_dir, src_tfrec_tbl, True) for fold_i, shard_tbl_fp in
-            enumerate(shard_tbls_fps)]
-    async_results = [pool.apply_async(create_shard_fold, job) for job in jobs]
-    pool.close()
+        # destination directory for the predicted shards
+        dest_tfrec_dir = data_dir / 'tfrecords' / 'predict'
+        dest_tfrec_dir.mkdir(exist_ok=True, parents=True)
 
-    # TCEs present in the fold TCE tables that were not present in the source TFRecords
-    tces_not_found_df = pd.concat([pd.DataFrame(async_result.get(), columns=['uid', 'target_id', 'tce_plnt_num'])
-                                   for async_result in async_results], ignore_index=True)
-    tces_not_found_df.to_csv(data_dir / 'tces_not_found_predict.csv', index=False)
+        pool = multiprocessing.Pool(processes=n_processes)
+        jobs = [(shard_tbl_fp, dest_tfrec_dir, fold_i, src_tfrec_dir, src_tfrec_tbl, True) for fold_i, shard_tbl_fp in
+                enumerate(shard_tbls_fps)]
+        async_results = [pool.apply_async(create_shard_fold, job) for job in jobs]
+        pool.close()
 
-    logger.info('Finished creating TFRecords shards for prediction.')
+        # TCEs present in the fold TCE tables that were not present in the source TFRecords
+        tces_not_found_df = pd.concat([pd.DataFrame(async_result.get(), columns=['uid', 'target_id', 'tce_plnt_num'])
+                                       for async_result in async_results], ignore_index=True)
+        tces_not_found_df.to_csv(data_dir / 'tces_not_found_predict.csv', index=False)
 
-    data_shards_fps = list(dest_tfrec_dir.iterdir())
-    data_shards = [data_shards_fp.name for data_shards_fp in data_shards_fps]
+        logger.info('Finished creating TFRecords shards for prediction.')
 
-    # assert len(data_dir.iterdir()) == num_folds
-    num_folds = len(data_shards_fps)
-    # assert num_unit_folds <= num_folds
-    folds_arr = np.arange(num_folds)
+        data_shards_fps = list(dest_tfrec_dir.iterdir())
+        data_shards = [data_shards_fp.name for data_shards_fp in data_shards_fps]
 
-    pred_shards = {'predict': data_shards}
-    np.save(data_dir / f'predict_shards.npy', pred_shards)
+        # assert len(data_dir.iterdir()) == num_folds
+        num_folds = len(data_shards_fps)
+        # assert num_unit_folds <= num_folds
+        folds_arr = np.arange(num_folds)
+
+        pred_shards = {'predict': data_shards}
+        np.save(data_dir / f'predict_shards.npy', pred_shards)
